Move utilities into OmniSharp.Utilities namespace
