+++ src/OmniSharp.Abstractions/Utilities/PlatformHelper.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 
-namespace OmniSharp
+namespace OmniSharp.Utilities
 {
     public static class PlatformHelper
     {
