+++ src/OmniSharp.DotNet/Tools/PackagesRestoreTool.cs
@@ -1,12 +1,11 @@
 using System;
 using System.Collections.Concurrent;
 using System.Diagnostics;
-using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.PlatformAbstractions;
 using OmniSharp.Services;
+using OmniSharp.Utilities;
 
 namespace OmniSharp.DotNet.Tools
 {
