+++ src/OmniSharp.Host/Middleware/LoggingMiddleware.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
+using OmniSharp.Utilities;
 
 namespace OmniSharp.Middleware
 {
